Save executed return code in callers $?
Save the return code of the execute command in $?. This perhaps isn't
the cleanest of solutions, to manipulate variables in caller scope.
However, it's similar to what bash does and it sort of makes sense, so
long the usage of the return code is a corner-case and not common
practice this should suffice.