execute: increase timeout in start
You might have such a slow connection to your target that a set timeout
of 1 second before we fail the start is very dangerous. Increase the
timeout to 10 seconds.

Observed in the wild over ssh connection with lots of starts generating
lots of output in a tight loop. Would be worse on slower serial
connections.