harness: add suite option translations
It is sometimes useful to be able to inject a small set of keywords in
the suite options parameters and have them translated to something
meaningful.

One such option is the directory path of the script the suite is
pointing out. Think a testcase that takes one or more file paths to
files containing test data. Such test would either need to have static
paths in the script or be unable to exist in a suite. This patch adds a
<base> keyword that can be used in the suite opts portion and will be
substituted with the path to the script the opts belong to before it is
passed to the test itself.