execute: allow for nested command execution
This is useful when spawning a shell. The user can start a shell
using cmd::start "/bin/bash" then execute commands inside the
shell while still being able to gracefully exit out of the shell and
capture it's return code. I wouldn't be surprised if there was more
usecases for this.