execute: fix simultaneous exec on different shells
Fix a bug where expecting on the active shell could be interrupted by
a command terminating on a different shell.

This bug was caused by the behaviour of expect_after. Expect_after
registers for a particular spawn (shell), it then looks at that
particular spawn. In our case it looks for a command termination. When
it gets the command termination it breaks the current expect block.

This resulted in any expect block checking all registered expect_after
calls sequentially and potentially breaking out of the expect block if
any of the running commands terminated.

Example:
9pm::shell::open "foo"
9pm::cmd::start "mycmd1"

9pm::shell::open "bar"
9pm::cmd::start "mycmd2"

Prior to this patch, an expect block issured here would look something
like this.

expect {
	"my explicit match" { puts "got my match" }
	-i $foo_spawn_id "foo-termination-checksum" {}
	-i $bar_spawn_id "bar-termination-checksum" {}
}

Meaning that if mycmd1 terminated before mycmd2 the expect clause
could break, even though mycmd1 terminated on an inactive spawn.

Fixes #9
Reported-by: Erik Hugne <erik.hugne@gmail.com>