config: remove 9pm::conf::get_req
After the change that no framework module calls 9pm::conf::* helpers
there are no longer any need to keep the 9pm::conf::get_req which
have been superseded by 9pm::helpers::dict::require as the function to
ensure a dict contains a key.

The user should not exploratory call 9pm::conf::get and check if the
return value is "" to determine if a key is present in the
configuration. If there is a need for this functionality a new function
should be added, 9pm::conf::exists or something similar. Therefor make
all calls on none existing keys fatal for 9pm::conf::get.

Remove 9pm::conf::get_req from both implementation and unit test.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>