9pm.py: use argparse for parsing cmdline
Move from getopt to argparse for parsing the cmdline.

Note: Previously, whenever 9pm displayed the help text (no matter if it
was because of a syntax error or if the user explicitly invoked '-h') it
would exit with exit code 1. Now, it will exit with code 0 if the user
used '-h' for help (which makes sense, considering it executed
successfully).

Signed-off-by: Andreas Bofjall <andreas@gazonk.org>