execute: Align start check with end check schema
Instead of depending on the start checksum to be printed on a separate
line align the check to match that of the end checksum. Print the start
checksum followed by a expression that will evaluate to a number ($$
evaluates to PID) that demonstrates the console is active.

This have no effect on spawns of ssh connections but greatly
improves reliability of spawns dealing with serial consoles.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>