execute: Correct assumption about shell behavior
There are little control of what is printed to the console between a
command inputted and the starts of the commands output. The 9pm library
assumes that the only thing printed between the two is a newline '\r\n'
as effect of the '\n' terminating the send operation.

This assumption have been observed to not always be true. Some systems
can and do inject extra characters between the two events which leads to
parts of 9pm to fail. Fix this by removing the assumption that the
output of a command inputted with send will always start with a '\r\n\'
sequence.

The most intrusive change is for ::9pm::cmd::execute where the start
checksum is enhanced to print the PID of the start checksum echo itself.
The idea is to have a more complex expression to match while only
depending on a string under our control.

For the unit-test the fix easier to see as it just removes the incorrect
assumption about shell behavior.

With these changes execute works and all unit tests pass.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>