execute: Add support for send_slow
Talking over serial consoles and sending long commands can exhaust the
targets serial FIFO if it's small enough. Expect have support to
mitigate such issues by the use the 'send_slow' variable and -s option
to send. Please see the expect manual for a full description of this
concept.

Add an optional argument to start '-slow' that if present configure and
use this mode when sending the command,

    ::9pm::cmd::start $longcmd -slow {10 .0001}

The idea of extending the arguments to start as -opt <value> pairs is so
that in the future other mitigation modes of send may be added without
conflicting.

If the -slow option is not given there is no change in behavior of the
current implementation.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>