helpers: add new function argument parser
Add a new parser that helps to pars non-positional function arguments.

Example:
foo {args} {
    set opts [9pm::misc::getopts $args]
    set foobar [dict get $opts "foobar"]
}
In this example, the only thing getopts does is to ensure args is a
dictionary.

It can also be used with default arguments:

foo {args} {
    set opts [9pm::misc::getopts $args "timeout" 10]

    expect {
    -timeout [dict get $opts "timeout"]
    ...
    }
}

In this scenario, the timeout is always set. Either to 10, if no
timeout argument is passed to foo or the passed value.