execute: add grace period and check after cmd abort
After a command is aborted, it might take some time for the console to
be ready. If a new command is sent before the console is ready, we run
into problems, which might make the console unresponsive.

To remedy this we first add a small sleep grace after aborting (10ms
currently). In a slow or loaded system this might still not be enough.
To catch such scenario this patch introduces a new abort reset check,
which runs the abort again after 100ms. This means sending ctrl+c
again and staring over.

Signed-off-by: Richard Alpe <rical@highwind.se>