execute: Allow caller to change start timeout
Allow the caller to specify how long to wait for a command to start
before it's declared a failure. This is useful when testing targets
where the console is in a state where it's slow to react, for example
when the system is waking up from suspend.

The default timeout is kept at 10 seconds and there is no functional
change.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>