output: Add wrapper to test and report result
The logging a test result as pass or fail based on a boolean variable is
not uncommon. I find myself often using a pattern like this,

    if {[sample_thermal_sensor]} {
        output::ok "Sample thermal sensor"
    } else {
        output::fail "Sample thermal sensor"
    }

To reduce the bloat in test-cases add a output::test helper to make this
use-case nicer to work with.

    output::test [sample_thermal_sensor] "Sample thermal sensor"

The helper also returns the test result so it can be used to clean up
after a failed test, in case the test script can recover and continue.

    if {![output::test [sample_thermal_sensor] "Sample thermal sensor"]} {
        emulate_thermal_sensor_values
    }

    use_thermal_sensor

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>