spawn: ssh: Add a new spawn centred around ssh
In addition to the standard shell spawn add a ssh spawn that directly
creates a ssh process. This could previously be  done by first creating
a local shell and then using the ssh helper to jump to the next host,
which still is valid. But for some use-cases it's nicer to spawn the ssh
process directly for safety.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>