sshshell: Add a new shell centred around ssh
In addition to the standard shell spawn add a sshshell that directly
spawns a ssh process instead of a local shell. This could previously be
done by first creating a local shell and then using the ssh helper to
jump to the next host, which still is valid. But for some use-cases it's
nicer to spawn the ssh process directly for safety.

This new spawn class is both directly useful and serve as an example on
how to further extend the spawn class to new uses-cases.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>