execute: Use int::cmd::get_last helper in capture()
The internal helper int::cmd::get_last is used in the rest of the file
to get the last command, use it in capture() as well instead of open
coding it. While at it align the variable name to 'last' to match the
pattern of the rest of the file.

There is no functional change.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>