Move tcl library code to lib_tcl subdirectory
9pm is getting some traction as a generic test executor. Running
arbitrary TAP executables in a controlled manner.

In this patch we move the TCL/Expect code to a subdirectory to
indicate that a user doesn't need to use the tcl "library" code.

The name lib_tcl should really be tcllib, but that can be confused
with the main upstream tcllib.

Signed-off-by: Richard Alpe <richard@bit42.se>