Use unique log dir for each 9pm execution
Redesign log_path handling so that each execution of 9pm.py creates
its own unique base dir.

For lib_tcl code, this simply moved the logfiles into a subdirectory
which is the date if it's set by 9pm.py or "standalone" if the lib_tcl
code is executed without 9pm.py as harness.

There's also a last symlink to indicate the last execution.

Signed-off-by: Richard Alpe <richard@bit42.se>