9pm: resolve <scratch> in suite option
Allow a suite to pass <scratch> to represent current SCRATCHDIR. This
is useful when the suite indicates the base path of something, such
as:

test.yaml:
opts:
- case : "create.py"
    - "output:<scratch>/result"

deploy.yaml:
opts:
- case : "create.py"
    - "output:/var/foobar/result"

In the first example, everything created by create.py will be removed
by 9pm at end of execution. In the second suite the same test is used
and the data is now stored persistently in /var/foobar/result.

Signed-off-by: Richard Alpe <richard@bit42.se>