lib_tcl: improve handling of extra ssh args from test
Unit testing in this area is unfortunately weak. Hopefully this causes
no regressions for users that may have worked around the previous
somewhat broken implementation.

The idea with the magic tcl variable "args" is that it allows the user
to pass an arbitrary amount of arguments at the end of the function
call. in 9pm::ssh::connect we expect this to look something like:

9pm::ssh::connect $node "-o StrictHostKeyChecking=no" "-o ..."

Signed-off-by: Richard Alpe <richard@bit42.se>