9pm.py: add suite settings that does nothing
Add a new "settings" option to a suite. This will allow the user to add various settings to a specific suite. Such as where to find the test-specification.

Currently it's only added to the data structure. I need to sort out
the horrible variable naming in parse_suite() before making it do
something.

Signed-off-by: Richard Alpe <richard@bit42.se>