9pm.py: redesign 9pm project config handling
It was a mistake to use -c (config), that config is intended solely
for the test case. It should really be named NINEPM_TEST_CONFIG and
not NINEPM_CONFIG. It's intended to make the test dynamic. For
example, the same test could be executed twice with different
configurations to test slightly different things.

The project config however is different. It shall only contain project
static data. Such as the project name. It's passed to test cases as
well, so a user could add additional top level data intended for the
test cases. Such as default user credentials for the project.

Signed-off-by: Richard Alpe <richard@bit42.se>