9pm.py: redesign handling of included test specifications
Prior to this commit, test specification where copied to the log
directory and renamed to there sha1sum. The idea was to avoid needing
the source code when generating the final .pdf report (not done by
9pm). However, test specifications rely on included data which also
has to be copied, making this solution messy and error prone.

In this commit we therefor abandon the idea of copying test
specifications to the log directory and instead add them with absolute
paths to the report.adoc (which is still located in the log
directory).

In addition to this we add a new project config variable
"PROJECT-TOPDIR" which is intended to specify the root path of the
relative test specification files.

Example:

From included test specification:
image::{topdoc}test/case/ietf_system/hostname/topology.svg[]

9pm cloned in: /home/user/foobar/9pm

PROJECT-TOPDOC: "../"

Resulting image include path (from generated report.adoc):
/home/user/foobar/9pm/../test/case/hostname/topology.svg

Signed-off-by: Richard Alpe <richard@bit42.se>