Add new self test framework (for 9pm.py)
This is a totally new 9pm.py self test suite. Its only purpose is to
ensure 9pm.py works as expected.

The old unit_tests tests both tcl and the harness together. The goal
now is to separate the tcl code from the python3 harness. This means
we need standalone tests for 9pm.py. Hence this commit.

Signed-off-by: Richard Alpe <richard@bit42.se>