Reformat test report output, blend spec + report in same section
Change disposition of the report from separate output and specification
sections, to a combined specification, result and output for each test.

This greatly improves the accessibility for the generated PDF, keeping
all relevant information in one place, and eliminating the unintuitive
link to the Test Specification from the output.  Links are problematic
in printed media.

Each test's heading now includes colored PASS/FAIL/SKIP which lines up
nicely in the generated Table of Contents, similar in fashion to how
write_report_result_tree() already works.  Several unicode symbols were
tested instead of the spelled-out result, e.g., ☑ but the result was
very difficult to read.

The call to write_report_result_tree() is temporarily dropped in this
commit, only to reintroduce it in the next in a new Test Info section
under the Test Summary.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>