report: adjust size of headings to avoid underline
Adjust size of heading slightly so it is evident that even a h4 is
distinguishable from bold text.

Drop underline for h3 since it does not play well with colored text.
The line under a colored word also becomes colored.

For links, which are removed in this patch series, let's use a clear
blue color so that users recognize them as links.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>