Fix annoying warnings when running in a container
This is a follow-up to 07f28a4 which added support for figuring out the
GIT revision when 9pm is a submodule in another proejct.

However, that project in turn may be a git worktree, where the actual
.git directory is elsewhere, and when said project is running its tests
from a Docker container, things get worse.  It is very likely that the
actual .git directory is unreachable from within the container.

This patch hides the following annoying warnings:

  warning, git command failed (Command '['git', '--git-dir', '/home/jocke/src/x-misc/test/9pm/../../../infix/.git/worktrees/x-misc/modules/9pm', 'rev-parse', 'HEAD']' returned non-zero exit status 128.)
9PM - Simplicity is the ultimate sophistication
  warning, git command failed (Command '['git', '--git-dir', '/home/jocke/src/infix/.git/worktrees/x-misc', 'rev-parse', 'HEAD']' returned non-zero exit status 128.)
  Testing Infix
  Running suite 0001 all.yaml

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>